+++ opencivicdata/tests/test_loaddivisions.py
@@ -1,6 +1,6 @@
 import pytest
 from django.core.management import call_command
-from opencivicdata.models import Division
+from opencivicdata.core.models import Division
 
 
 @pytest.mark.django_db